wave 3 completed, integrated application view
